// frontend/src/api/schema.ts
export interface components {
    schemas: {
// ...
        MovementType: "purchase" | "sale" | "dividend" | "interest" | "fee" | "deposit" | "withdrawal" | "transfer_out" | "transfer_in" | "principal_repayment";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}